fix
fix: 修正中止/重啟邏輯;

+++ index.js
@@ -82,8 +82,8 @@ client.once("ready", () => {
 
 // 監聽 SIGTERM 訊號（Render 停止服務時會發送此信號）
 let isStoppingBot = false;
-/* process.on('SIGTERM', async () => {
-    // 如果 !stopTheDiscordBot 則跳過重啟
+process.on('SIGTERM', async () => {
+    // !stopTheDiscordBot 則跳過重啟
     if (isStoppingBot) return;
 
     console.log('[INFO]已收到 SIGTERM 訊號，正在開始重啟程序...');
@@ -104,7 +104,7 @@ let isStoppingBot = false;
         console.error('[ERROR]無法觸發部署', err);
     }
     process.exit(0);
-}); */
+});
 
 // 定義 Slash 命令列表
 const theCommands = [
@@ -116,6 +116,9 @@ const theCommands = [
 
 // 監聽 Slash Command
 client.on("interactionCreate", async (interaction) => {
+    // !stopTheDiscordBot 後進入假眠
+    if (isStoppingBot) return;
+
     if (!interaction.isCommand()) return;
 
     if (interaction.commandName === "help") {
@@ -125,18 +128,24 @@ client.on("interactionCreate", async (interaction) => {
 
 // 監聽 keywords
 client.on("messageCreate", async (message) => {
+    // !stopTheDiscordBot 後進入假眠
+    if (isStoppingBot) return;
+
     // 忽略 Bot 自己的訊息
     if (message.author.bot) return;
 
     const content = message.content;
 
-    // 捕獲中止命令
+    // 捕獲中止命令 !stopTheDiscordBot
     if (content.includes("!stopTheDiscordBot") && message.member.roles.cache.has(process.env.ADMIN_ROLE_ID)) {
         isStoppingBot = true;
         await message.reply("おやすみなさい。");
         console.log("[INFO]theDiscordBot 停止中...");
         client.destroy(); // 停止 Discord Bot
-        process.exit(0); // 終止程式
+        server.close(() => {
+            console.log("[INFO]Web Server 已關閉");
+        });
+        rewturn; // 用 process.exit(0) 會被render重啟
     }
 
     // 處理符合關鍵字的命令
